+++ contracts/adapter/chainlink/ChainlinkAdapterInternal.sol
@@ -29,10 +29,10 @@ abstract contract ChainlinkAdapterInternal is
 
     /// @dev If a fresh price is unavailable the adapter will wait the duration of
     ///      MAX_DELAY before returning the stale price
-    uint32 internal constant MAX_DELAY = 12 hours;
+    uint256 internal constant MAX_DELAY = 12 hours;
     /// @dev If the difference between target and last update is greater than the
     ///      PRICE_STALE_THRESHOLD, the price is considered stale
-    uint32 internal constant PRICE_STALE_THRESHOLD = 25 hours;
+    uint256 internal constant PRICE_STALE_THRESHOLD = 25 hours;
 
     constructor(
         address _wrappedNativeToken,
@@ -42,7 +42,7 @@ abstract contract ChainlinkAdapterInternal is
     function _quoteFrom(
         address tokenIn,
         address tokenOut,
-        uint32 target
+        uint256 target
     ) internal view returns (UD60x18) {
         (
             PricingPath path,
